+++ examples/ai-core/src/generate-text/anthropic-custom-fetch.ts
@@ -3,13 +3,14 @@ import { generateText } from 'ai';
 import 'dotenv/config';
 
 const anthropic = createAnthropic({
-  // example fetch wrapper that logs the URL:
+  // example fetch wrapper that logs the input to the API call:
   fetch: async (url, options) => {
-    console.log(`Fetching ${url}`);
-    const result = await fetch(url, options);
-    console.log(`Fetched ${url}`);
-    console.log();
-    return result;
+    console.log('URL', url);
+    console.log('Headers', JSON.stringify(options!.headers, null, 2));
+    console.log(
+      `Body ${JSON.stringify(JSON.parse(options!.body! as string), null, 2)}`,
+    );
+    return await fetch(url, options);
   },
 });