+++ examples/next-openai/app/use-chat-reasoning/page.tsx
@@ -17,26 +17,40 @@ export default function Chat() {
   });
 
   return (
-    <div className="flex flex-col w-full max-w-md py-24 mx-auto stretch">
+    <div className="flex flex-col w-full max-w-2xl py-24 mx-auto stretch">
       {messages.map(message => (
-        <div key={message.id} className="whitespace-pre-wrap">
-          {message.role === 'user' ? 'User: ' : 'AI: '}
-          {message.parts.map((part, index) => {
-            if (part.type === 'text') {
-              return <div key={index}>{part.text}</div>;
-            }
+        <div
+          key={message.id}
+          className="flex gap-4 pb-4 mb-6 border-b border-gray-100 last:border-0"
+        >
+          <div className="font-medium min-w-[50px]">
+            {message.role === 'user' ? 'User:' : 'AI:'}
+          </div>
+          <div className="flex-1">
+            {message.parts.map((part, index) => {
+              if (part.type === 'text') {
+                return (
+                  <pre
+                    key={index}
+                    className="max-w-full overflow-x-auto break-words whitespace-pre-wrap"
+                  >
+                    {part.text}
+                  </pre>
+                );
+              }
 
-            if (part.type === 'reasoning') {
-              return (
-                <pre
-                  key={index}
-                  className="italic text-gray-500 whitespace-pre-wrap"
-                >
-                  {part.reasoning}
-                </pre>
-              );
-            }
-          })}
+              if (part.type === 'reasoning') {
+                return (
+                  <pre
+                    key={index}
+                    className="max-w-full mb-4 overflow-x-auto italic text-gray-500 break-words whitespace-pre-wrap"
+                  >
+                    {part.reasoning}
+                  </pre>
+                );
+              }
+            })}
+          </div>
         </div>
       ))}
 
@@ -68,7 +82,7 @@ export default function Chat() {
 
       <form onSubmit={handleSubmit}>
         <input
-          className="fixed bottom-0 w-full max-w-md p-2 mb-8 border border-gray-300 rounded shadow-xl"
+          className="fixed bottom-0 w-full max-w-2xl p-2 mb-8 border border-gray-300 rounded shadow-xl"
           value={input}
           placeholder="Say something..."
           onChange={handleInputChange}