+++ packages/google-vertex/src/anthropic/google-vertex-anthropic-messages-settings.ts
@@ -2,6 +2,7 @@ import { AnthropicMessagesSettings } from '@ai-sdk/anthropic/internal';
 
 // https://cloud.google.com/vertex-ai/generative-ai/docs/partner-models/use-claude
 export type GoogleVertexAnthropicMessagesModelId =
+  | 'claude-3-7-sonnet@20250219'
   | 'claude-3-5-sonnet-v2@20241022'
   | 'claude-3-5-haiku@20241022'
   | 'claude-3-5-sonnet@20240620'