+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,40 @@
 
 ### Patch Changes
 
+- Updated dependencies [ddf9740]
+  - ai@4.1.46
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [93bd5a0]
+  - ai@4.1.45
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [f8e7df2]
+  - ai@4.1.44
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [ef2e23b]
+  - ai@4.1.43
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.1.42
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@4.1.41
 
 ## 0.0.1