+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/openai-compatible
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [ddf9740]
+  - @ai-sdk/provider@1.0.9
+  - @ai-sdk/provider-utils@2.1.10
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [2761f06]
+  - @ai-sdk/provider@1.0.8
+  - @ai-sdk/provider-utils@2.1.9
+
 ## 0.1.10
 
 ### Patch Changes