+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @ai-sdk/valibot
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [ddf9740]
+  - ai@4.1.46
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [93bd5a0]
+  - ai@4.1.45
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [f8e7df2]
+  - ai@4.1.44
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [ef2e23b]
+  - ai@4.1.43
+
+## 0.0.9
+
+### Patch Changes
+
+- ai@4.1.42
+
 ## 0.0.8
 
 ### Patch Changes